+++ packages/solid/rollup.config.js
@@ -3,4 +3,5 @@ import withSolid from 'rollup-preset-solid';
 export default withSolid({
   input: 'src/index.ts',
   targets: ['esm', 'cjs'],
+  external: ['valibot', 'zod'],
 });