+++ packages/solid/src/adapters/index.ts
@@ -1,2 +1,4 @@
+export * from './valiField';
+export * from './valiForm';
 export * from './zodField';
 export * from './zodForm';